Move out delivery_result to separate file
